made util functions and structured functions
