+++ packages/cbioportal-utils/src/model/Exon.ts
@@ -1,0 +1,1 @@
+export type ExonDatum = { rank: number; length: number; start: number };
