+++ packages/react-mutation-mapper/src/component/track/TrackPanel.tsx
@@ -12,7 +12,7 @@ import MutationMapperStore from '../../model/MutationMapperStore';
 import { TrackName, TrackVisibility } from './TrackSelector';
 import HotspotTrack from './HotspotTrack';
 import OncoKbTrack from './OncoKbTrack';
-import ExonNumTrack from './ExonTrack';
+import ExonTrack from './ExonTrack';
 import PtmTrack from './PtmTrack';
 
 import './defaultTrackTooltipTable.scss';
@@ -36,7 +36,7 @@ export default class TrackPanel extends React.Component<TrackPanelProps, {}> {
             TrackName.CancerHotspots,
             TrackName.OncoKB,
             TrackName.dbPTM,
-            TrackName.ExonNum,
+            TrackName.Exon,
         ],
     };
 
@@ -86,10 +86,10 @@ export default class TrackPanel extends React.Component<TrackPanelProps, {}> {
                 TrackName.UniprotPTM,
                 PtmSource.Uniprot
             ),
-            [TrackName.ExonNum]:
+            [TrackName.Exon]:
                 !this.props.trackVisibility ||
-                this.props.trackVisibility[TrackName.ExonNum] === 'visible' ? (
-                    <ExonNumTrack
+                this.props.trackVisibility[TrackName.Exon] === 'visible' ? (
+                    <ExonTrack
                         store={this.props.store}
                         dataStore={this.props.store.dataStore}
                         width={this.props.geneWidth}
