+++ packages/react-mutation-mapper/src/util/TrackUtils.ts
@@ -7,6 +7,6 @@ export function initDefaultTrackVisibility(): TrackVisibility {
         [TrackName.dbPTM]: 'hidden',
         [TrackName.UniprotPTM]: 'hidden',
         [TrackName.PDB]: 'hidden',
-        [TrackName.ExonNum]: 'hidden',
+        [TrackName.Exon]: 'hidden',
     };
 }
