+++ src/shared/components/mutationMapper/MutationMapper.tsx
@@ -329,7 +329,7 @@ export default class MutationMapper<
             [TrackName.dbPTM]: dbPtmDataStatus,
             [TrackName.UniprotPTM]: uniprotPtmDataStatus,
             [TrackName.PDB]: alignmentDataStatus,
-            [TrackName.ExonNum]: 'complete',
+            [TrackName.Exon]: 'complete',
         };
     }
 
@@ -547,7 +547,7 @@ export default class MutationMapper<
         } else {
             tracks.push(TrackName.dbPTM);
         }
-        tracks.push(TrackName.ExonNum);
+        tracks.push(TrackName.Exon);
         tracks.push(TrackName.PDB);
 
         return tracks;
